Uncomment section

+++ scripts/cleanup-dcp1/cleanup.py
@@ -33,9 +33,9 @@ if __name__ == '__main__':
             
             logger.info(f'project {project_uuid} is a DCP1 project')
 
-            # if not input('Would you like to continue with fixing the terra area for this submission? (Y/n)').lower() == 'y':
-            #     skipped.append((uuid, 'Manually skipped'))
-            #     continue
+            if not input('Would you like to continue with fixing the terra area for this submission? (Y/n)').lower() == 'y':
+                skipped.append((uuid, 'Manually skipped'))
+                continue
 
             utils.run_command(f'gsutil -m rm -r {GCP_BUCKET}/{project_uuid}/metadata/sequence_file')
             utils.run_command(f'gsutil -m rm -r {GCP_BUCKET}/{project_uuid}/metadata/supplementary_file')
